Fix return type address space checking.
solves: `func-resource-result-simple.slang`, related to #4291

when getting the address space of a function return we now check the return address of the inst if the return type is a non pointer type.

Note: for `case kIROp_Return` if the type is a non pointer I believe we should just be early leaving, but for correctness `getFuncResultAddrSpace` was modified since address space of a pointer like type should be interpreted as `AddressSpace::Global` and not `AddressSpace::Generic`.